Packaged setup screens, partially implemented database restructure.
