+++ app/src/main/AndroidManifest.xml
@@ -9,19 +9,19 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SetupVerification"></activity>
-        <activity android:name=".SetupComplete" />
-        <activity android:name=".SetupHighBloodSugar" />
-        <activity android:name=".SetupWaitTimer" />
-        <activity android:name=".SetupAddCaregiver" />
-        <activity android:name=".SetupUserInfo2" />
+        <activity android:name=".Setup.SetupVerification"></activity>
+        <activity android:name=".Setup.SetupComplete" />
+        <activity android:name=".Setup.SetupHighBloodSugar" />
+        <activity android:name=".Setup.SetupWaitTimer" />
+        <activity android:name=".Setup.SetupAddCaregiver" />
+        <activity android:name=".Setup.SetupUserInfo2" />
         <activity android:name=".Dashboard" />
-        <activity android:name=".SetupCaregiverInfo" />
-        <activity android:name=".SetupUserInfo" />
-        <activity android:name=".SetupSecondaryTreatment" />
-        <activity android:name=".SetupPrimaryTreatment" />
-        <activity android:name=".SetupBloodSugar" />
-        <activity android:name=".SetupUserType" />
+        <activity android:name=".Setup.SetupCaregiverInfo" />
+        <activity android:name=".Setup.SetupUserInfo" />
+        <activity android:name=".Setup.SetupSecondaryTreatment" />
+        <activity android:name=".Setup.SetupPrimaryTreatment" />
+        <activity android:name=".Setup.SetupBloodSugar" />
+        <activity android:name=".Setup.SetupUserType" />
         <activity android:name=".LoginActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
