+++ app/src/main/java/com/example/sixthsenseapp/Dashboard.java
@@ -1,13 +1,38 @@
 package com.example.sixthsenseapp;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.Toast;
+
+import com.google.firebase.auth.FirebaseAuth;
 
 public class Dashboard extends AppCompatActivity {
+
+    private Button logOutButton;
+    private FirebaseAuth mAuth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_dashboard);
+
+        logOutButton = (Button) findViewById(R.id.logoutButton);
+
+        mAuth = LoginActivity.getFirebaseAuth();
+
+        logOutButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                mAuth.getInstance().signOut();
+
+                Intent intent = new Intent(Dashboard.this, MainActivity.class);
+                startActivity(intent);
+            }
+        });
     }
 }
