+++ app/src/main/java/com/example/sixthsenseapp/Setup/SetupAddCaregiver.java
@@ -1,4 +1,4 @@
-package com.example.sixthsenseapp;
+package com.example.sixthsenseapp.Setup;
 
 import android.content.Intent;
 
@@ -7,6 +7,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+
+import com.example.sixthsenseapp.R;
 
 public class SetupAddCaregiver extends AppCompatActivity {
