+++ app/src/main/java/com/example/sixthsenseapp/Setup/SetupBloodSugar.java
@@ -1,4 +1,4 @@
-package com.example.sixthsenseapp;
+package com.example.sixthsenseapp.Setup;
 
 import android.content.Intent;
 import android.graphics.Color;
@@ -10,6 +10,8 @@ import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.sixthsenseapp.R;
 
 import java.text.DecimalFormat;
