+++ app/src/main/java/com/example/sixthsenseapp/Setup/SetupPrimaryTreatment.java
@@ -1,18 +1,16 @@
-package com.example.sixthsenseapp;
+package com.example.sixthsenseapp.Setup;
 
 import android.content.Intent;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
-import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.view.View;
-import android.widget.Button;
-import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
-import android.widget.ToggleButton;
+
+import com.example.sixthsenseapp.R;
 
 public class SetupPrimaryTreatment extends AppCompatActivity {
