+++ app/src/main/java/com/example/sixthsenseapp/Setup/SetupSecondaryTreatment.java
@@ -1,17 +1,17 @@
-package com.example.sixthsenseapp;
+package com.example.sixthsenseapp.Setup;
 
 import android.content.Intent;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.media.Image;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.View;
-import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.sixthsenseapp.R;
 
 public class SetupSecondaryTreatment extends AppCompatActivity {
