+++ app/src/main/java/com/example/sixthsenseapp/Setup/SetupUserInfo2.java
@@ -1,59 +1,67 @@
-package com.example.sixthsenseapp;
+package com.example.sixthsenseapp.Setup;
 
 import android.content.Intent;
-import android.graphics.Color;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.View;
+import android.widget.CheckBox;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-public class SetupUserInfo extends AppCompatActivity {
+import com.example.sixthsenseapp.R;
+
+public class SetupUserInfo2 extends AppCompatActivity {
 
     private ImageView backgroundImage;
     private ImageButton nextButton;
     private ImageButton backButton;
-    private EditText firstNameField;
-    private EditText lastNameField;
-    private EditText emailAddressField;
-    private EditText passwordField;
-    private EditText retypePasswordField;
-    private EditText phoneNumberField;
+    private EditText dateOfBirthField;
+    private EditText gpNameField;
+    private EditText gpNumberField;
+    private EditText emergencyNameField;
+    private EditText emergencyNumberField;
     private TextView errorMessage;
-    private static String firstName = "";
-    private static String lastName = "";
-    private static String emailAddress = "";
-    private static String password = "";
-    private static String retypePassword = "";
-    private static String phoneNumber = "";
+    private CheckBox addCaregiver;
+    private static String dateOfBirth = "";
+    private static String gpName = "";
+    private static String gpNumber = "";
+    private static String emergencyName = "";
+    private static String emergencyNumber = "";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_setup_user_info);
+        setContentView(R.layout.activity_setup_user_info2);
 
         backgroundImage = (ImageView) findViewById(R.id.backgroundImage);
         nextButton = (ImageButton) findViewById(R.id.nextButton);
         backButton = (ImageButton) findViewById(R.id.backButton);
-        firstNameField = (EditText) findViewById(R.id.firstNameField);
-        lastNameField = (EditText) findViewById(R.id.lastNameField);
-        emailAddressField = (EditText) findViewById(R.id.emailField);
-        passwordField = (EditText) findViewById(R.id.passwordField);
-        retypePasswordField = (EditText) findViewById(R.id.retypePasswordField);
-        phoneNumberField = (EditText) findViewById(R.id.phoneNumberField);
+        dateOfBirthField = (EditText) findViewById(R.id.dateOfBirthField);
+        gpNameField = (EditText) findViewById(R.id.gpNameField);
+        gpNumberField = (EditText) findViewById(R.id.gpNumberField);
+        emergencyNameField = (EditText) findViewById(R.id.emergencyNameField);
+        emergencyNumberField = (EditText) findViewById(R.id.emergencyNumberField);
         errorMessage = (TextView) findViewById(R.id.errorMessage);
+        addCaregiver = (CheckBox) findViewById(R.id.addCaregiver);
 
         int imageResource = getResources().getIdentifier("@drawable/loginbackground", null, this.getPackageName());
         backgroundImage.setImageResource(imageResource);
 
         nextButton.setEnabled(false);
 
-        firstNameField.addTextChangedListener(new TextWatcher() {
+        if(SetupUserType.getUserType().equals("User")){
+            addCaregiver.setVisibility(View.VISIBLE);
+        }
+        else if(SetupUserType.getUserType().equals("Caregiver")){
+            addCaregiver.setVisibility(View.INVISIBLE);
+        }
+
+        dateOfBirthField.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
 
@@ -70,7 +78,7 @@ public class SetupUserInfo extends AppCompatActivity {
             }
         });
 
-        lastNameField.addTextChangedListener(new TextWatcher() {
+        gpNameField.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
 
@@ -87,7 +95,7 @@ public class SetupUserInfo extends AppCompatActivity {
             }
         });
 
-        emailAddressField.addTextChangedListener(new TextWatcher() {
+        gpNumberField.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
 
@@ -104,7 +112,7 @@ public class SetupUserInfo extends AppCompatActivity {
             }
         });
 
-        passwordField.addTextChangedListener(new TextWatcher() {
+        emergencyNameField.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
 
@@ -121,24 +129,7 @@ public class SetupUserInfo extends AppCompatActivity {
             }
         });
 
-        retypePasswordField.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-                checkTextFields();
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-
-            }
-        });
-
-        phoneNumberField.addTextChangedListener(new TextWatcher() {
+        emergencyNumberField.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence s, int start, int count, int after) {
 
@@ -158,27 +149,25 @@ public class SetupUserInfo extends AppCompatActivity {
         nextButton.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View view){
+                dateOfBirth = dateOfBirthField.getText().toString();
+                gpName = gpNameField.getText().toString();
+                gpNumber = gpNumberField.getText().toString();
+                emergencyName = emergencyNameField.getText().toString();
+                emergencyNumber = emergencyNumberField.getText().toString();
 
-                firstName = firstNameField.getText().toString();
-                lastName = lastNameField.getText().toString();
-                emailAddress = emailAddressField.getText().toString();
-                password = passwordField.getText().toString();
-                retypePassword = retypePasswordField.getText().toString();
-                phoneNumber = phoneNumberField.getText().toString();
-
-                if(password.length() >= 8){
-                    if(password.equals(retypePassword)){
-                        Intent intent = new Intent(SetupUserInfo.this, SetupUserInfo2.class);
+                if(SetupUserType.getUserType().equals("User")){
+                    if(addCaregiver.isChecked()){
+                        Intent intent = new Intent(SetupUserInfo2.this, SetupAddCaregiver.class);
                         startActivity(intent);
                     }
                     else{
-                        errorMessage.setTextColor(Color.RED);
-                        errorMessage.setText("Passwords do not match!");
+                        Intent intent = new Intent(SetupUserInfo2.this, SetupBloodSugar.class);
+                        startActivity(intent);
                     }
                 }
-                else{
-                    errorMessage.setTextColor(Color.RED);
-                    errorMessage.setText("Password must be at least 8 characters!");
+                else if(SetupUserType.getUserType().equals("Caregiver")){
+                    Intent intent = new Intent(SetupUserInfo2.this, SetupCaregiverInfo.class);
+                    startActivity(intent);
                 }
             }
         });
@@ -186,15 +175,15 @@ public class SetupUserInfo extends AppCompatActivity {
         backButton.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View view){
-                Intent intent = new Intent(SetupUserInfo.this, SetupUserType.class);
+                Intent intent = new Intent(SetupUserInfo2.this, SetupUserInfo.class);
                 startActivity(intent);
             }
         });
     }
 
     private void checkTextFields(){
-        if((!firstNameField.getText().toString().isEmpty()) && (!lastNameField.getText().toString().isEmpty()) && (!emailAddressField.getText().toString().isEmpty())
-                && (!passwordField.getText().toString().isEmpty()) && (!retypePasswordField.getText().toString().isEmpty()) && (!phoneNumberField.getText().toString().isEmpty())){
+        if((!dateOfBirthField.getText().toString().isEmpty()) && (!gpNameField.getText().toString().isEmpty()) && (!gpNumberField.getText().toString().isEmpty())
+                && (!emergencyNameField.getText().toString().isEmpty()) && (!emergencyNumberField.getText().toString().isEmpty())){
             nextButton.setEnabled(true);
         }
         else{
@@ -202,23 +191,23 @@ public class SetupUserInfo extends AppCompatActivity {
         }
     }
 
-    public static String getFirstName(){
-        return firstName;
+    public static String getDateOfBirth(){
+        return dateOfBirth;
     }
 
-    public static String getLastName(){
-        return lastName;
+    public static String getGpName(){
+        return gpName;
     }
 
-    public static String getEmailAddress(){
-        return emailAddress;
+    public static String getGpNumber(){
+        return gpNumber;
     }
 
-    public static String getPassword(){
-        return password;
+    public static String getEmergencyName(){
+        return emergencyName;
     }
 
-    public static String getPhoneNumber(){
-        return phoneNumber;
+    public static String getEmergencyNumber(){
+        return emergencyNumber;
     }
 }
