+++ app/src/main/java/com/example/sixthsenseapp/Setup/SetupUserType.java
@@ -1,4 +1,4 @@
-package com.example.sixthsenseapp;
+package com.example.sixthsenseapp.Setup;
 
 import android.content.Intent;
 import androidx.appcompat.app.AppCompatActivity;
@@ -6,6 +6,9 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+
+import com.example.sixthsenseapp.MainActivity;
+import com.example.sixthsenseapp.R;
 
 public class SetupUserType extends AppCompatActivity {
