+++ app/src/main/res/layout/activity_setup_blood_sugar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupBloodSugar">
+    tools:context=".Setup.SetupBloodSugar">
 
     <ImageView
         android:id="@+id/backgroundImage"
