+++ app/src/main/res/layout/activity_setup_caregiver_info.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupCaregiverInfo">
+    tools:context=".Setup.SetupCaregiverInfo">
 
     <ImageView
         android:id="@+id/backgroundImage"
