+++ app/src/main/res/layout/activity_setup_complete.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupComplete">
+    tools:context=".Setup.SetupComplete">
 
     <ImageView
         android:id="@+id/backgroundImage"
