+++ app/src/main/res/layout/activity_setup_high_blood_sugar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupHighBloodSugar">
+    tools:context=".Setup.SetupHighBloodSugar">
 
     <ImageView
         android:id="@+id/backgroundImage"
