+++ app/src/main/res/layout/activity_setup_primary_treatment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupPrimaryTreatment">
+    tools:context=".Setup.SetupPrimaryTreatment">
 
     <ImageView
         android:id="@+id/backgroundImage"
