+++ app/src/main/res/layout/activity_setup_secondary_treatment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupSecondaryTreatment">
+    tools:context=".Setup.SetupSecondaryTreatment">
 
     <ImageView
         android:id="@+id/backgroundImage"
