+++ app/src/main/res/layout/activity_setup_user_info2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupUserInfo2">
+    tools:context=".Setup.SetupUserInfo2">
 
     <ImageView
         android:id="@+id/backgroundImage"
