+++ app/src/main/res/layout/activity_setup_user_type.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupUserType">
+    tools:context=".Setup.SetupUserType">
 
     <ImageView
         android:id="@+id/backgroundImage"
