+++ app/src/main/res/layout/activity_setup_wait_timer.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupWaitTimer">
+    tools:context=".Setup.SetupWaitTimer">
 
     <ImageView
         android:id="@+id/backgroundImage"
